moving qml helpers into helpers.py,removing calls in __init__
