+++ nexus_constructor/qml_models/instrument_model.py
@@ -9,7 +9,7 @@ from nexus_constructor.data_model import (
     Rotation,
     Translation,
 )
-from nexus_constructor.qml_models import change_value, generate_unique_name
+from nexus_constructor.qml_models.helpers import change_value, generate_unique_name
 from nexus_constructor.qml_models.transform_model import TransformationModel
 from nexus_constructor.off_renderer import OffMesh
 from PySide2.QtCore import Qt, QAbstractListModel, QModelIndex, Signal, Slot
